fix: resolve all linting errors for production code
- Fix type errors in search_unified.go:
  - Add missing HybridScore field to HybridSearchResult struct
  - Add missing QueryEmbedding field to HybridSearchRequest struct
  - Fix type conversions between float32/float64
  - Add missing getModelFamily helper function

- Fix embedding adapter errors:
  - Replace undefined SearchServiceV2 with AdvancedSearchService
  - Fix type mismatches in threshold/similarity conversions
  - Handle nullable ContextID properly

- Fix error handling:
  - Add proper error checks for rows.Close() and resp.Body.Close()
  - Add logger to SQLRepository for better error reporting

- Fix static analysis issues:
  - Remove empty branch in search_unified.go
  - Add context key types in websocket handlers to avoid collisions
  - Fix ineffective break statements with labeled loops
  - Remove redundant nil check for map length
  - Add nolint comments for backwards compatibility tests

- Clean up unused code:
  - Remove unused calculateCapabilityScore function
  - Remove unused calculateAgentPerformance functions
  - Remove outdated test files for removed services

- Update dependencies:
  - Import observability package where needed
  - Fix method signatures to match updated interfaces

All modules now pass linting checks successfully.

+++ apps/rest-api/internal/api/embedding_api.go
@@ -197,7 +197,7 @@ func (api *EmbeddingAPI) crossModelSearch(c *gin.Context) {
 		TenantID:       req.TenantID,
 		ContextID:      req.ContextID,
 		Limit:          req.Limit,
-		Threshold:      req.MinSimilarity,
+		Threshold:      float64(req.MinSimilarity),
 	}
 
 	// Add metadata filter if provided
